requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
deepl==1.19.1
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
gevent==24.10.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.1
openai==1.51.2
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.1.0
